# Conflicts:
#	server/bundles/io.cloudbeaver.model/src/io/cloudbeaver/model/config/CBServerConfig.java
#	server/bundles/io.cloudbeaver.server/src/io/cloudbeaver/WebServiceUtils.java
#	server/bundles/io.cloudbeaver.service.admin/src/io/cloudbeaver/service/admin/impl/WebServiceAdmin.java
#	server/bundles/io.cloudbeaver.service.security/src/io/cloudbeaver/service/security/EmbeddedSecurityControllerFactory.java

+++ webapp/packages/plugin-data-viewer/src/TableViewer/TableViewer.tsx
@@ -23,16 +23,19 @@ import {
   useSplitUserState,
   useTranslate,
 } from '@cloudbeaver/core-blocks';
-import type { IDataContext } from '@cloudbeaver/core-data-context';
 import { useService } from '@cloudbeaver/core-di';
 import { ResultDataFormat } from '@cloudbeaver/core-sdk';
+import { CaptureView } from '@cloudbeaver/core-view';
 
-import { ResultSetConstraintAction } from '../DatabaseDataModel/Actions/ResultSet/ResultSetConstraintAction';
+import { DatabaseDataConstraintAction } from '../DatabaseDataModel/Actions/DatabaseDataConstraintAction';
+import { IDatabaseDataOptions } from '../DatabaseDataModel/IDatabaseDataOptions';
 import { DataPresentationService, DataPresentationType } from '../DataPresentationService';
+import { isResultSetDataModel } from '../ResultSet/isResultSetDataModel';
+import { DataPresentation } from './DataPresentation';
+import { DataViewerViewService } from './DataViewerViewService';
 import type { IDataTableActionsPrivate } from './IDataTableActions';
 import { TableError } from './TableError';
 import { TableFooter } from './TableFooter/TableFooter';
-import { TableGrid } from './TableGrid';
 import { TableHeader } from './TableHeader/TableHeader';
 import { TablePresentationBar } from './TablePresentationBar/TablePresentationBar';
 import { TableToolsPanel } from './TableToolsPanel';
@@ -46,7 +49,6 @@ export interface TableViewerProps {
   valuePresentationId: string | null | undefined;
   /** Display data in simple mode, some features will be hidden or disabled */
   simple?: boolean;
-  context?: IDataContext;
   className?: string;
   onPresentationChange: (id: string) => void;
   onValuePresentationChange: (id: string | null) => void;
@@ -54,32 +56,28 @@ export interface TableViewerProps {
 
 export const TableViewer = observer<TableViewerProps, HTMLDivElement>(
   forwardRef(function TableViewer(
-    {
-      tableId,
-      resultIndex = 0,
-      presentationId,
-      valuePresentationId,
-      simple = false,
-      context,
-      className,
-      onPresentationChange,
-      onValuePresentationChange,
-    },
+    { tableId, resultIndex = 0, presentationId, valuePresentationId, simple = false, className, onPresentationChange, onValuePresentationChange },
     ref,
   ) {
     const translate = useTranslate();
     const styles = useS(style);
+    const dataViewerView = useService(DataViewerViewService);
     const dataPresentationService = useService(DataPresentationService);
     const tableViewerStorageService = useService(TableViewerStorageService);
     const dataModel = tableViewerStorageService.get(tableId);
-    const result = dataModel?.getResult(resultIndex);
+    const result = dataModel?.source.getResult(resultIndex);
     const loading = dataModel?.isLoading() ?? true;
     const dataFormat = result?.dataFormat || ResultDataFormat.Resultset;
     const splitState = useSplitUserState('table-viewer');
 
     const localActions = useObjectRef({
       clearConstraints() {
-        const constraints = dataModel?.source.tryGetAction(resultIndex, ResultSetConstraintAction);
+        const unknownModel = dataModel as any;
+        if (!isResultSetDataModel<IDatabaseDataOptions>(unknownModel)) {
+          return;
+        }
+
+        const constraints = unknownModel?.source.tryGetAction(resultIndex, DatabaseDataConstraintAction);
 
         if (constraints) {
           constraints.deleteAll();
@@ -155,9 +153,7 @@ export const TableViewer = observer<TableViewerProps, HTMLDivElement>(
       ['setPresentation', 'setValuePresentation', 'switchValuePresentation', 'closeValuePresentation'],
     );
 
-    const needRefresh = getComputed(
-      () => dataModel?.source.error === null && dataModel.source.results.length === 0 && dataModel.source.outdated && dataModel.source.isLoadable(),
-    );
+    const needRefresh = getComputed(() => !dataModel?.isDisabled(resultIndex) && dataModel?.source.isOutdated() && dataModel.source.isLoadable());
 
     useEffect(() => {
       if (needRefresh) {
@@ -180,7 +176,7 @@ export const TableViewer = observer<TableViewerProps, HTMLDivElement>(
     // }, [dataFormat]);
 
     if (!dataModel) {
-      return <Loader />;
+      return <TextPlaceholder>{translate('plugin_data_viewer_no_available_presentation')}</TextPlaceholder>;
     }
 
     const presentation = dataPresentationService.getSupported(DataPresentationType.main, dataFormat, presentationId, dataModel, resultIndex);
@@ -204,88 +200,90 @@ export const TableViewer = observer<TableViewerProps, HTMLDivElement>(
       !simple;
 
     return (
-      <div ref={ref} className={s(styles, { tableViewer: true }, className)}>
-        <div className={s(styles, { tableContent: true })}>
-          {!isStatistics && (
-            <TablePresentationBar
-              className={s(styles, { tablePresentationBar: true })}
-              type={DataPresentationType.main}
-              presentationId={presentation.id}
-              dataFormat={dataFormat}
-              supportedDataFormat={dataModel.supportedDataFormats}
-              model={dataModel}
-              resultIndex={resultIndex}
-              onPresentationChange={dataTableActions.setPresentation}
-            />
-          )}
-          <div className={s(styles, { tableData: true })}>
-            <TableHeader model={dataModel} resultIndex={resultIndex} simple={simple} />
-            <Split
-              className={s(styles, { split: true, disabled: !valuePanelDisplayed })}
-              {...splitState}
-              sticky={30}
-              mode={valuePanelDisplayed ? splitState.mode : 'minimize'}
-              disable={!valuePanelDisplayed}
-              keepRatio
-              disableAutoMargin
-            >
-              <Pane className={s(styles, { pane: true })}>
-                <div className={s(styles, { paneContent: true, grid: true })}>
-                  <Loader className={s(styles, { loader: true })} suspense>
-                    <TableGrid
-                      model={dataModel}
-                      actions={dataTableActions}
-                      dataFormat={dataFormat}
-                      presentation={presentation}
-                      resultIndex={resultIndex}
-                      simple={simple}
-                      isStatistics={isStatistics}
-                    />
-                  </Loader>
-                  <TableError model={dataModel} loading={loading} />
-                  <Loader
-                    loading={loading}
-                    cancelDisabled={!dataModel.source.canCancel}
-                    overlay={overlay}
-                    onCancel={() => dataModel.source.cancel()}
-                  />
-                </div>
-              </Pane>
-              <ResizerControls />
-              <Pane className={s(styles, { pane: true })} basis="30%" main>
-                <Loader className={s(styles, { loader: true })} suspense>
-                  <div className={s(styles, { paneContent: true })}>
-                    {resultExist && (
-                      <TableToolsPanel
+      <CaptureView className={s(styles, { captureView: true })} view={dataViewerView}>
+        <div ref={ref} className={s(styles, { tableViewer: true }, className)}>
+          <div className={s(styles, { tableContent: true })}>
+            {!isStatistics && (
+              <TablePresentationBar
+                className={s(styles, { tablePresentationBar: true })}
+                type={DataPresentationType.main}
+                presentationId={presentation.id}
+                dataFormat={dataFormat}
+                supportedDataFormat={dataModel.supportedDataFormats}
+                model={dataModel}
+                resultIndex={resultIndex}
+                onPresentationChange={dataTableActions.setPresentation}
+              />
+            )}
+            <div className={s(styles, { tableData: true })}>
+              <TableHeader model={dataModel} resultIndex={resultIndex} simple={simple} />
+              <Split
+                className={s(styles, { split: true, disabled: !valuePanelDisplayed })}
+                {...splitState}
+                sticky={30}
+                mode={valuePanelDisplayed ? splitState.mode : 'minimize'}
+                disable={!valuePanelDisplayed}
+                keepRatio
+                disableAutoMargin
+              >
+                <Pane className={s(styles, { pane: true })}>
+                  <div className={s(styles, { paneContent: true, grid: true })}>
+                    <Loader className={s(styles, { loader: true })} suspense>
+                      <DataPresentation
                         model={dataModel}
                         actions={dataTableActions}
                         dataFormat={dataFormat}
-                        presentation={valuePresentation}
+                        presentation={presentation}
                         resultIndex={resultIndex}
                         simple={simple}
+                        isStatistics={!!isStatistics}
                       />
-                    )}
+                    </Loader>
+                    <TableError model={dataModel} loading={loading} />
+                    <Loader
+                      loading={loading}
+                      cancelDisabled={!dataModel.source.canCancel}
+                      overlay={overlay}
+                      onCancel={() => dataModel.source.cancel()}
+                    />
                   </div>
-                </Loader>
-              </Pane>
-            </Split>
+                </Pane>
+                <ResizerControls />
+                <Pane className={s(styles, { pane: true })} basis="30%" main>
+                  <Loader className={s(styles, { loader: true })} suspense>
+                    <div className={s(styles, { paneContent: true })}>
+                      {resultExist && (
+                        <TableToolsPanel
+                          model={dataModel}
+                          actions={dataTableActions}
+                          dataFormat={dataFormat}
+                          presentation={valuePresentation}
+                          resultIndex={resultIndex}
+                          simple={simple}
+                        />
+                      )}
+                    </div>
+                  </Loader>
+                </Pane>
+              </Split>
+            </div>
+            {!simple && !isStatistics && (
+              <TablePresentationBar
+                type={DataPresentationType.toolsPanel}
+                className={s(styles, { tablePresentationBar: true })}
+                presentationId={valuePresentationId ?? null}
+                dataFormat={dataFormat}
+                supportedDataFormat={[dataFormat]}
+                model={dataModel}
+                resultIndex={resultIndex}
+                onPresentationChange={dataTableActions.setValuePresentation}
+                onClose={dataTableActions.closeValuePresentation}
+              />
+            )}
           </div>
-          {!simple && !isStatistics && (
-            <TablePresentationBar
-              type={DataPresentationType.toolsPanel}
-              className={s(styles, { tablePresentationBar: true })}
-              presentationId={valuePresentationId ?? null}
-              dataFormat={dataFormat}
-              supportedDataFormat={[dataFormat]}
-              model={dataModel}
-              resultIndex={resultIndex}
-              onPresentationChange={dataTableActions.setValuePresentation}
-              onClose={dataTableActions.closeValuePresentation}
-            />
-          )}
+          <TableFooter model={dataModel} resultIndex={resultIndex} simple={simple} />
         </div>
-        <TableFooter model={dataModel} resultIndex={resultIndex} simple={simple} context={context} />
-      </div>
+      </CaptureView>
     );
   }),
 );
